Add a 'universal' background colour fix
This means that the Dracula backgorund colour is applied based on meaning, not specific elements. Hopefully this won't have any unintended side-effects (it should effectively replace the default backgroud colour, wherever it's used, with the Dracula one).
This fixes the new Server Home Page background.

This commit also includes a fix to make sure the correct gradient colour is applied to the 'fade out' at the end of channel topics.